+++ CHANGELOG.md
@@ -11,14 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [GH#183](https://github.com/jolicode/automapper/pull/183) Ability to change reload strategy from AutoMapper::create()
 - [GH#193](https://github.com/jolicode/automapper/pull/193) add icon to symfony profiler
 
+### Changed
+- [GH#186](https://github.com/jolicode/automapper/pull/186) Optimize creation from constructor
+
 ### Fixed
 - [GH#184](https://github.com/jolicode/automapper/pull/184) Fix error when mapping from stdClass to constructor with nullable/optional arguments
 - [GH#185](https://github.com/jolicode/automapper/pull/185) Fix constructor with default parameter array does not work with constructor_arguments context
 - [GH#187](https://github.com/jolicode/automapper/pull/187) Fix regression after [GH#184](https://github.com/jolicode/automapper/pull/184)
 - [GH#192](https://github.com/jolicode/automapper/pull/192) Fix source and context not passed to callable transformer
-
-### Changed
-- [GH#186](https://github.com/jolicode/automapper/pull/186) Optimize creation from constructor
 
 ## [9.1.2] - 2024-09-03
 ### Fixed