+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [9.2.0] - 2024-11-19
 ### Added
 - [GH#180](https://github.com/jolicode/automapper/pull/180) Add configuration to generate code with strict types
 - [GH#183](https://github.com/jolicode/automapper/pull/183) Ability to change reload strategy from AutoMapper::create()
@@ -346,6 +348,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [AutoMapper] [GH#179](https://github.com/janephp/janephp/pull/179) Fixing incompatible changes in Symfony 5.0
 
 [Unreleased]: https://github.com/jolicode/automapper/compare/9.1.2...HEAD
+[9.2.0]: https://github.com/janephp/janephp/compare/9.1.1...9.1.2
 [9.1.2]: https://github.com/janephp/janephp/compare/9.1.1...9.1.2
 [9.1.1]: https://github.com/janephp/janephp/compare/9.1.0...9.1.1
 [9.1.0]: https://github.com/janephp/janephp/compare/9.0.2...9.1.0